+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/sink/concurrent/ConcurrentRdaSink.java
@@ -3,6 +3,7 @@ package gov.cms.bfd.pipeline.rda.grpc.sink.concurrent;
 import com.google.common.annotations.VisibleForTesting;
 import gov.cms.bfd.pipeline.rda.grpc.ProcessingException;
 import gov.cms.bfd.pipeline.rda.grpc.RdaSink;
+import gov.cms.bfd.pipeline.rda.grpc.sink.reactive.ReactiveRdaSink;
 import java.io.IOException;
 import java.time.Duration;
 import java.util.Collection;
