+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/sink/reactive/StringPartitioner.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.pipeline.rda.grpc.sink.concurrent;
+package gov.cms.bfd.pipeline.rda.grpc.sink.reactive;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.hash.HashFunction;
